fix silly warnings and typos
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@3868 599ab11a-0111-0410-8a50-d7a5b0953570

+++ librapi2/src/config/config.c
@@ -257,7 +257,7 @@ int getConfigInt (struct configFile *cfg, char *section, char *key)
 	return atoi (configString);
 }
 
-int getConfigDouble (struct configFile *cfg, char *section, char *key)
+double getConfigDouble (struct configFile *cfg, char *section, char *key)
 {
 	char *configString;
 	if ( (configString=getConfigString(cfg, section, key))==NULL) {
